mod: update TabularDataset.
